+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -11533,7 +11533,7 @@ $CONN_Completers = {
             ($_ -eq "Update-CONNInstanceStorageConfig/ResourceType")
         }
         {
-            $v = "AGENT_EVENTS","CALL_RECORDINGS","CHAT_TRANSCRIPTS","CONTACT_TRACE_RECORDS","MEDIA_STREAMS","SCHEDULED_REPORTS"
+            $v = "AGENT_EVENTS","CALL_RECORDINGS","CHAT_TRANSCRIPTS","CONTACT_TRACE_RECORDS","MEDIA_STREAMS","REAL_TIME_CONTACT_ANALYSIS_SEGMENTS","SCHEDULED_REPORTS"
             break
         }
 
@@ -13404,6 +13404,13 @@ $DGURU_Completers = {
             break
         }
 
+        # Amazon.DevOpsGuru.EventSourceOptInStatus
+        "Update-DGURUEventSourcesConfig/EventSources_AmazonCodeGuruProfiler_Status"
+        {
+            $v = "DISABLED","ENABLED"
+            break
+        }
+
         # Amazon.DevOpsGuru.InsightType
         {
             ($_ -eq "Get-DGURUOrganizationInsightList/StatusFilter_Any_Type") -Or
@@ -13458,6 +13465,7 @@ $DGURU_Completers = {
 
 $DGURU_map = @{
     "Action"=@("Update-DGURUResourceCollection")
+    "EventSources_AmazonCodeGuruProfiler_Status"=@("Update-DGURUEventSourcesConfig")
     "Filters_DataSource"=@("Get-DGURUEventList")
     "Filters_EventClass"=@("Get-DGURUEventList")
     "Locale"=@("Get-DGURURecommendationList")
@@ -13523,6 +13531,7 @@ $DGURU_SelectMap = @{
                "Get-DGURUAccountHealth",
                "Get-DGURUAccountOverview",
                "Get-DGURUAnomaly",
+               "Get-DGURUEventSourcesConfig",
                "Get-DGURUFeedback",
                "Get-DGURUInsight",
                "Get-DGURUOrganizationHealth",
@@ -13543,6 +13552,7 @@ $DGURU_SelectMap = @{
                "Search-DGURUInsight",
                "Search-DGURUOrganizationInsight",
                "Start-DGURUCostEstimation",
+               "Update-DGURUEventSourcesConfig",
                "Update-DGURUResourceCollection",
                "Update-DGURUServiceIntegration")
 }