+++ modules/AWSPowerShell/Cmdlets/Macie/Basic/Update-MACS3Resource-Cmdlet.cs
@@ -28,11 +28,12 @@ using Amazon.Macie.Model;
 namespace Amazon.PowerShell.Cmdlets.MAC
 {
     /// <summary>
-    /// Updates the classification types for the specified S3 resources. If memberAccountId
-    /// isn't specified, the action updates the classification types of the S3 resources associated
-    /// with Amazon Macie Classic for the current Macie Classic administrator account. If
-    /// memberAccountId is specified, the action updates the classification types of the S3
-    /// resources associated with Macie Classic for the specified member account.
+    /// (Discontinued) Updates the classification types for the specified S3 resources. If
+    /// <code>memberAccountId</code> isn't specified, the action updates the classification
+    /// types of the S3 resources associated with Amazon Macie Classic for the current Macie
+    /// Classic administrator account. If <code>memberAccountId</code> is specified, the action
+    /// updates the classification types of the S3 resources associated with Macie Classic
+    /// for the specified member account.
     /// </summary>
     [Cmdlet("Update", "MACS3Resource", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.Macie.Model.FailedS3Resource")]
@@ -47,8 +48,8 @@ namespace Amazon.PowerShell.Cmdlets.MAC
         #region Parameter MemberAccountId
         /// <summary>
         /// <para>
-        /// <para>The AWS ID of the Amazon Macie Classic member account whose S3 resources' classification
-        /// types you want to update. </para>
+        /// <para>(Discontinued) The Amazon Web Services account ID of the Amazon Macie Classic member
+        /// account whose S3 resources' classification types you want to update.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
@@ -58,7 +59,7 @@ namespace Amazon.PowerShell.Cmdlets.MAC
         #region Parameter S3ResourcesUpdate
         /// <summary>
         /// <para>
-        /// <para>The S3 resources whose classification types you want to update.</para>
+        /// <para>(Discontinued) The S3 resources whose classification types you want to update.</para>
         /// </para>
         /// </summary>
         #if !MODULAR